+++ element/table_of_contents_element.js
@@ -93,7 +93,7 @@ TableOfContents.setAttribute('intersection-class');
  *
  * @author   Jelle De Loecker   <jelle@elevenways.be>
  * @since    0.1.2
- * @version  0.1.2
+ * @version  0.2.0
  */
 TableOfContents.setProperty(function entries() {
 
@@ -114,7 +114,10 @@ TableOfContents.setProperty(function entries() {
 	}
 
 	if (wrapper) {
-		let elements = wrapper.querySelectorAll(this.elements_selector || 'h1,h2'),
+
+		let current_level = null,
+		    last_entry,
+		    elements = wrapper.querySelectorAll(this.elements_selector || 'h1,h2'),
 			element,
 			title,
 			i;
@@ -133,7 +136,9 @@ TableOfContents.setProperty(function entries() {
 				}
 			}
 
-			let title_element;
+			let title_element,
+			    starts_level,
+				ends_level;
 
 			if (this.title_selector) {
 				title_element = element.querySelector(this.title_selector);
@@ -141,6 +146,25 @@ TableOfContents.setProperty(function entries() {
 
 			if (!title_element) {
 				title_element = element;
+			}
+
+			if (element.nodeName[0] == 'H' && isFinite(element.nodeName[1])) {
+				let heading_level = +element.nodeName[1];
+
+				if (current_level == null) {
+					current_level = heading_level;
+				} else if (heading_level > current_level) {
+					current_level++;
+					starts_level = true;
+				} else if (heading_level == current_level && last_entry) {
+					last_entry.starts_level = false;
+				} else if (heading_level < current_level) {
+					current_level--;
+
+					if (last_entry) {
+						last_entry.ends_level = true;
+					}
+				}
 			}
 
 			title = (title_element.toc_title || title_element.textContent || '').trim();
@@ -152,11 +176,74 @@ TableOfContents.setProperty(function entries() {
 				continue;
 			}
 
-			result.push({
-				id      : element.id,
-				title   : title,
-				element : element,
-			});
+			last_entry = {
+				id           : element.id,
+				level        : current_level,
+				title        : title,
+				element      : element,
+				starts_level : starts_level,
+				ends_level   : ends_level,
+			};
+
+			result.push(last_entry);
+		}
+
+		let ended_level = false,
+		    entries = result,
+		    current_branch,
+			current_nodes = [];
+
+		result = current_nodes;
+		last_entry = null;
+
+		for (let entry of entries) {
+
+			if (ended_level) {
+
+				while (entry.level >= current_branch.level) {
+					current_branch = current_branch.parent;
+					current_nodes = current_branch?.children || result;
+
+					if (!current_branch) {
+						current_branch = result[result.length - 1];
+						break;
+					}
+				}
+
+				ended_level = false;
+			}
+
+			if (!current_branch) {
+				current_branch = entry;
+			} else {
+				entry.parent = current_branch;
+			}
+
+			if (entry.starts_level) {
+
+				current_branch = last_entry;
+
+				if (!current_branch.children) {
+					current_branch.children = [];
+				}
+
+				current_nodes = current_branch.children;
+			}
+			
+			current_nodes.push(entry);
+
+			if (entry.ends_level) {
+				ended_level = true;
+				current_branch = entry.parent;
+				current_nodes = current_branch.children || result;
+
+				if (!current_branch) {
+					current_branch = result[result.length - 1];
+					break;
+				}
+			}
+
+			last_entry = entry;
 		}
 	}